실시간 리뷰 분석 API 수정

+++ lib/services/review_service.dart
@@ -1,6 +1,7 @@
 // services/review_service.dart
 import 'dart:convert';
 import 'dart:io';
+import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:http/http.dart' as http;
 
@@ -60,7 +61,14 @@ class ReviewService {
     return response.statusCode == 201;
   }
 
-  Future<bool> deleteReview(String reviewId, String token) async {
+  Future<bool> deleteReview(String? reviewId, String? token) async {
+    if (reviewId == null ||
+        reviewId.isEmpty ||
+        token == null ||
+        token.isEmpty) {
+      return false;
+    }
+
     final url = Uri.parse('$baseUrl/api/review/$reviewId');
     final response = await http.delete(
       url,
@@ -86,6 +94,48 @@ class ReviewService {
       }),
     );
 
+    debugPrint('📡 PATCH 상태 코드: ${response.statusCode}');
+    debugPrint('📨 응답 본문: ${response.body}');
+
     return response.statusCode == 200;
+  }
+
+  Future<List<Map<String, String>>> getReviewsByUser(
+      String? token, String? userId) async {
+    if (token == null || token.isEmpty || userId == null || userId.isEmpty) {
+      return [];
+    }
+
+    final url = Uri.parse('$baseUrl/api/review/user/$userId');
+    final response = await http.get(
+      url,
+      headers: {
+        'Authorization': 'Bearer $token',
+        'Content-Type': 'application/json',
+      },
+    );
+    if (response.statusCode == 200) {
+      final Map<String, dynamic> data = json.decode(response.body);
+      final List<dynamic> reviews = data['reviews'];
+
+      return reviews.map<Map<String, String>>((review) {
+        final content = review['content'] ?? '';
+        final location = review['location'];
+        final id = review['_id'];
+        final locationName =
+            location is Map<String, dynamic> ? location['title'] ?? '' : '';
+        final locationId =
+            location is Map<String, dynamic> ? location['_id'] ?? '' : '';
+
+        return {
+          'id': id,
+          'content': content,
+          'location': locationName,
+          'locationId': locationId,
+        };
+      }).toList();
+    } else {
+      throw Exception('리뷰 조회 실패');
+    }
   }
 }
